+++ packages/swap/src/payment.ts
@@ -1,4 +1,5 @@
 import { Contract } from '@bitcoin-computer/lib'
+import { getTestRev } from './utils'
 
 export class Payment extends Contract {
   _id: string
@@ -9,6 +10,27 @@ export class Payment extends Contract {
 
   constructor(owner: string, _amount: number) {
     super({ _owners: [owner], _amount })
+  }
+
+  transfer(to: string) {
+    this._owners = [to]
+  }
+}
+
+export class PaymentMock {
+  _id: string
+  _rev: string
+  _root: string
+  _amount: number
+  _owners: string[]
+
+  constructor(owner: string, amount: number) {
+    const r = Math.floor(Math.random() * 1000)
+    this._id = getTestRev(0, r)
+    this._rev = getTestRev(0, r)
+    this._root = getTestRev(0, r)
+    this._owners = [owner]
+    this._amount = amount
   }
 
   transfer(to: string) {
