+++ packages/swap/test/ord-sale.test.ts
@@ -6,87 +6,13 @@ import chaiMatchPattern from 'chai-match-pattern'
 import { Computer } from '@bitcoin-computer/lib'
 import { Transaction } from '@bitcoin-computer/nakamotojs'
 import { NFT } from '@bitcoin-computer/TBC721/src/nft'
-import { Sale, Valuable } from '../src/ord-sale'
-import { Payment } from '../src/payment'
+import { Sale } from '../src/ord-sale'
+import { Payment, PaymentMock } from '../src/payment'
+import { RLTC, meta } from '../src/utils'
+import { Valuable, ValuableMock } from '../src/valuable'
 
 chai.use(chaiMatchPattern)
 const _ = chaiMatchPattern.getLodashModule()
-
-const randomPublicKey = '023e21361b53bb2e625cc1f41d18b35ae882e88d8d107df1c3711fa8bc54db8fed'
-
-const RLTC: {
-  network: 'regtest'
-  chain: 'LTC'
-  url: string
-} = {
-  network: 'regtest',
-  chain: 'LTC',
-  url: 'http://localhost:1031',
-}
-
-const meta = {
-  _id: _.isString,
-  _rev: _.isString,
-  _root: _.isString,
-  _owners: _.isArray,
-  _amount: _.isNumber,
-}
-
-export function getTestTxId(i = 0): string {
-  if (i === 0) return 'mock:0000000000000000000000000000000000000000000000000000000000000000'
-  if (i === 1) return 'mock:1111111111111111111111111111111111111111111111111111111111111111'
-  if (i === 2) return 'mock:2222222222222222222222222222222222222222222222222222222222222222'
-  if (i === 3) return 'mock:3333333333333333333333333333333333333333333333333333333333333333'
-  if (i === 4) return 'mock:4444444444444444444444444444444444444444444444444444444444444444'
-  if (i === 5) return 'mock:5555555555555555555555555555555555555555555555555555555555555555'
-  throw new Error('getTestTxId is only defined for parameters smaller than 6.')
-}
-
-export function getTestRev(txId = 0, outNum = 0): string {
-  return `${getTestTxId(txId)}:${outNum}`
-}
-
-class PaymentMock {
-  _id: string
-  _rev: string
-  _root: string
-  _amount: number
-  _owners: string[]
-
-  constructor(owner: string, amount: number) {
-    const r = Math.floor(Math.random() * 1000)
-    this._id = getTestRev(0, r)
-    this._rev = getTestRev(0, r)
-    this._root = getTestRev(0, r)
-    this._owners = [owner]
-    this._amount = amount
-  }
-
-  transfer(to: string) {
-    this._owners = [to]
-  }
-}
-
-class ValuableMock {
-  _id: string
-  _rev: string
-  _root: string
-  _amount: number
-  _owners: string[]
-
-  constructor() {
-    const r = Math.floor(Math.random() * 100000)
-    this._id = getTestRev(1, r)
-    this._rev = getTestRev(1, r)
-    this._root = getTestRev(1, r)
-    this._owners = [randomPublicKey]
-    this._amount = 7860
-  }
-
-  setAmount(amount: number) {
-    this._amount = amount
-  }
-}
 
 describe('Ord Sale', () => {
   let tx: any
